colorapi/requirements.txt
numpy==1.19.4
Flask==1.0.2
Pillow==7.1.1
scikit_learn==0.23.2
